Wed, Nov 22, 2023, 4:03 PM +01:00

+++ interpreter/interpreter.py
@@ -5,11 +5,10 @@ z = int(z)
 
 match y:
     case "+":
-        print(x + z)
+        print(float(x + z))
     case "-":
-        print(x - z)
+        print(float(x - z))
     case "*":
-        print(x * z)
+        print(float(x * z))
     case "/":
-        
-        print(x / z)
+        print(float(x / z))
